Fix Skia build on macOS
The x11 and wayland features of Skia enable the build of x11/wayland specific code to determine
`skia_safe::gpu::gl::Interface`.
That code does not compile on macOS.

In theory they could be masked to only be enabled on linux, but even there we don't need it, because
we provide the `skia_safe::gpu::gl::Interface` instance.