Disable the Skia binary cache download mechanism
It's sort of dangerous to download binaries off of GitHub and unfortunately it
breaks with the Yocto build, because it brings in
build dependencies (ring) that in turn try to invoke the host assembler by calling HOST_CC via cc.rs

Unfortunately Corrosion sets HOST_CC to CC, which Yocto sets to be the host compiler.

This is a workaround. Revisit this :-)